remove unused lifetime
