+++ ffi/rodbus-ffi/src/client.rs
@@ -1,5 +1,5 @@
 use crate::ffi;
-use rodbus::client::{ReconnectStrategy, WriteMultiple};
+use rodbus::client::WriteMultiple;
 use rodbus::AddressRange;
 use std::path::Path;
 use std::time::Duration;
@@ -270,10 +270,4 @@ pub(crate) unsafe fn client_channel_set_decode_level(
         .runtime
         .spawn(channel.inner.set_decode_level(level.into()))?;
     Ok(())
-}
-
-impl From<ffi::RetryStrategy> for Box<dyn ReconnectStrategy + Send> {
-    fn from(from: ffi::RetryStrategy) -> Self {
-        rodbus::client::doubling_reconnect_strategy(from.min_delay(), from.max_delay())
-    }
 }
